Brayden changes + rename to third_party
