+++ pkg/client/client_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/sonr-io/sonr/pkg/crypto/mpc"
-	"github.com/sonr-io/sonr/thirdparty/types/common"
+	"github.com/sonr-io/sonr/third_party/types/common"
 
 	"github.com/stretchr/testify/assert"
 )
