+++ pkg/motor/bucket_test.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/sonr-io/sonr/thirdparty/types/common"
-	mt "github.com/sonr-io/sonr/thirdparty/types/motor"
+	"github.com/sonr-io/sonr/third_party/types/common"
+	mt "github.com/sonr-io/sonr/third_party/types/motor"
 	"github.com/sonr-io/sonr/x/bucket/types"
 	"github.com/stretchr/testify/assert"
 )
