+++ pkg/motor/create_account.go
@@ -8,9 +8,9 @@ import (
 	"github.com/sonr-io/sonr/pkg/client"
 	"github.com/sonr-io/sonr/pkg/crypto/mpc"
 	"github.com/sonr-io/sonr/pkg/did"
-	mt "github.com/sonr-io/sonr/thirdparty/types/motor"
 	"github.com/sonr-io/sonr/pkg/tx"
 	"github.com/sonr-io/sonr/pkg/vault"
+	mt "github.com/sonr-io/sonr/third_party/types/motor"
 	rt "github.com/sonr-io/sonr/x/registry/types"
 )
