+++ pkg/motor/object_builder_test.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/sonr-io/sonr/thirdparty/types/common"
-	mt "github.com/sonr-io/sonr/thirdparty/types/motor"
+	"github.com/sonr-io/sonr/third_party/types/common"
+	mt "github.com/sonr-io/sonr/third_party/types/motor"
 	"github.com/stretchr/testify/assert"
 )
