+++ pkg/motor/resource.go
@@ -13,24 +13,24 @@ import (
 )
 
 type motorResources struct {
-	config       *client.Client
-	shell        *shell.Shell
-	whatIsStore  map[string]*st.WhatIs
-	whereIsStore map[string]*bt.WhereIs
-	schemaStore  map[string]*st.SchemaDefinition
-	bucketStore  map[string]bucket.Bucket
+	config            *client.Client
+	shell             *shell.Shell
+	whatIsStore       map[string]*st.WhatIs
+	whereIsStore      map[string]*bt.WhereIs
+	schemaStore       map[string]*st.SchemaDefinition
+	bucketStore       map[string]bucket.Bucket
 }
 
 func newMotorResources(
 	config *client.Client,
 	shell *shell.Shell) *motorResources {
 	return &motorResources{
-		config:       config,
-		shell:        shell,
-		bucketStore:  make(map[string]bucket.Bucket),
-		whatIsStore:  make(map[string]*st.WhatIs),
-		whereIsStore: make(map[string]*bt.WhereIs),
-		schemaStore:  make(map[string]*st.SchemaDefinition),
+		config:            config,
+		shell:             shell,
+		bucketStore:       make(map[string]bucket.Bucket),
+		whatIsStore:       make(map[string]*st.WhatIs),
+		whereIsStore:      make(map[string]*bt.WhereIs),
+		schemaStore:       make(map[string]*st.SchemaDefinition),
 	}
 }
