+++ pkg/motor/x/exchange/protocol.go
@@ -9,7 +9,7 @@ import (
 	"github.com/patrickmn/go-cache"
 	"github.com/sonr-io/sonr/pkg/config"
 	"github.com/sonr-io/sonr/pkg/host"
-	st "github.com/sonr-io/sonr/thirdparty/types/service"
+	st "github.com/sonr-io/sonr/third_party/types/service"
 )
 
 type ExchangeProtocol struct {
