+++ pkg/motor/x/transmit/file.go
@@ -16,7 +16,7 @@ import (
 	"time"
 
 	"github.com/gabriel-vasile/mimetype"
-	st "github.com/sonr-io/sonr/thirdparty/types/service"
+	st "github.com/sonr-io/sonr/third_party/types/service"
 )
 
 var (
