+++ proto/common/v1/info.proto
@@ -3,7 +3,7 @@ syntax = "proto3";
 // Package common defines commonly used types agnostic to the node role on the Sonr network.
 package sonrio.common.v1;
 
-option go_package = "github.com/sonr-io/sonr/thirdparty/types/common";
+option go_package = "github.com/sonr-io/sonr/third_party/types/common";
 
 // Basic Info Sent to Peers to Establish Connections
 message Peer {
