+++ proto/motor/v1/response.proto
@@ -8,7 +8,7 @@ import "registry/who_is.proto";
 import "schema/v1/what_is.proto";
 import "common/v1/info.proto";
 
-option go_package = "github.com/sonr-io/sonr/thirdparty/types/motor";
+option go_package = "github.com/sonr-io/sonr/third_party/types/motor";
 
 
 message InitializeResponse {
