+++ proto/service/v1/transmit.proto
@@ -4,7 +4,7 @@ syntax = "proto3";
 package sonrio.motor.service.v1;
 import "common/v1/info.proto";
 
-option go_package = "github.com/sonr-io/sonr/thirdparty/types/service";
+option go_package = "github.com/sonr-io/sonr/third_party/types/service";
 
 // Direction is the direction of a stream.
 enum Direction {
